+++ Multisig/Assets.xcassets/Tab Bar/ico-app-settings.imageset/ico-app-settings.pdf
