+++ Multisig/Assets.xcassets/Tab Bar/ico-safe-settings.imageset/ico-safe-settings.pdf
